Refresh latest builds for all packages
It should be better to refresh latest tagged builds for all packages,
not just tracked ones.  Firstly, it is faster as there is no need to
use multicall.  Secondly, knowing latest bulid for untracked packages
makes it possible to submit scratch build as soon as tracked flag is
set to true, without having to poll koji.