Refactor plugin structure
Previous state: separate plugins for backend and frontend
Current state: plugins for whole koschei with submodules for backend and frontend.

The directory structure:
plugins/$name/$endpoint
where
name = name of the plugin, directory must be a python package - must contain
       __init__.py (it can contain common parts or be empty)
endpoint = "backend" or "frontend". It may be a python module (e.g. backend.py)
           or python package (backend directory with __init__.py). May contain
           just one endpoint or both.
the endpoint module/package is imported on startup when the plugin is enabled.

"backend" endpoint may also contain "services" directory, which will be scanned
for services the same way as in main koschei package.