[resolver] Avoid race conditions
When sqla is building query results and sees an object that it already
has in its session, it returns that object ignoring the actual row from
query result. With that it tries to emulate repeatable read level of
isolation. But in koschei, we sometimes need read committed isolation
level in order to achieve mutual exclusion of side-effects (fedmsg,
build scheduling). Expiring the objects in resolver forces their refresh,
avoiding race conditions with scheduler (like described in #100 and #158).