Query SRPM for requested architectures
Recently, Koji started lying about arch of noarch builds of packages
that have arch exclusions, such as [0] (is noarch, but reports arch=i386).
This attempts to use data from SRPM headers to determine whether
a package is noarch and on which architectures it should be built.
Also skips packages for which the arch list was empty (no allowed arch).

[0] https://koji.fedoraproject.org/koji/taskinfo?taskID=19927957