[views] Fix skip_resolution being set to False when editing groups
HTML POST request don't contain values of unchecked checkboxes, so
WTForms treats all not present BooleanFields as False.
Which causes skip_resolution to be set to False when editing package
groups, because they share a form. I introduced a custom field that uses
additional hidden field to determine checkbox presence. I could have as
well just split the forms, but I think this is more systematic solution,
as it also handles the (very rare, but possible) case when koschei is
upgraded and someone submits a form that gained additional field during
the upgrade.