Remove collection.dest_tag
Scheduler and polling are using dest_tag for querying SRPMs and real builds.
For Rawhide it doesn't really matter, but for stable Fedoras, it can
cause different effects:
- The builds are later moved by bodhi into different tags, so they disappear
- Build overrides don't have effect

Moreover, the dest_tag is often wrong, as it reflects the configuration
by the time the collection was setup, which may differ from what is in
koji now.