[views] Highlight changed version component
Highlight the leftmost components of dependency changes which changed.
Example:
1.[6]-3.fc28 -> 1.[7]-1.fc29 # [] denotes bold

This is to give better visual aid when scanning through large lists of
dependency changes. Especially after mass-rebuilds, when everything was
release bumped, but there may be packages which had been updated to new
upstream versions, and thus have higher probability of affecting the
build state.

Note that it works on string representation, so technically it's not
RPM-correct in cases like 7 -> 07. I think such cases are so rare that
it's not worth the complexity a RPM-correct solution would have had.