[resolver] Query the build group on secondary
Primary and secondary should have the same build group. But they don't
have the same repos which we now use for getting the build group at
a particular point in time. So it needs to query the koji which owns the
repo, which is secondary. It is also more consistent - both repo and
build group are queried on secondary.