Make distgit group refresh script use extras JSON
Pagure broke its API, starting to require pagination.  Unfortunately
Pagure pagination is prone to race conditions and returns inconsistent
data. Not to mention it's slow and complex to use.  As such it can't
be relied on by Koschei.

The koschei-refresh-distgit-group script is changed to use data
produced extras JSON produced by pagure-dist-git instead of Pagure
API.